spring-02-library-mongo/src/main/java/ru/otus/spring/library/service/impl/IncSequenceGeneratorService.java
package ru.otus.spring.library.service.impl;

import lombok.RequiredArgsConstructor;
import org.springframework.data.mongodb.core.MongoOperations;
import org.springframework.data.mongodb.core.query.Update;
import org.springframework.stereotype.Service;
import ru.otus.spring.library.domain.DatabaseSequence;
import ru.otus.spring.library.service.SequenceGeneratorService;

import java.util.Objects;

import static org.springframework.data.mongodb.core.FindAndModifyOptions.options;
import static org.springframework.data.mongodb.core.query.Criteria.where;
import static org.springframework.data.mongodb.core.query.Query.query;

@Service
@RequiredArgsConstructor
public class IncSequenceGeneratorService implements SequenceGeneratorService<Integer> {

    private final MongoOperations mongoOperations;

    @Override
    public Integer generateSequence(String seqName) {
        DatabaseSequence counter = mongoOperations.findAndModify(
                query(where("_id").is(seqName)),
                new Update().inc(DatabaseSequence.Fields.seq, 1), options().returnNew(true).upsert(true),
                DatabaseSequence.class
        );
        return !Objects.isNull(counter) ? counter.getSeq() : 1;
    }
}
